refactoring datatypes

+++ trackcomment/src/main/java/dwsc/trackcomment/repository/CommentRepository.java
@@ -6,12 +6,12 @@ import org.springframework.data.repository.CrudRepository;
 
 import dwsc.trackcomment.model.Comment;
 
-public interface CommentRepository extends CrudRepository<Comment, Long> {
+public interface CommentRepository extends CrudRepository<Comment, Integer> {
 
 	<S extends Comment> S save(S comment);
 
 	List<Comment> findAll();
 
-	List<Comment> findByTrackid(Long trackid);
+	List<Comment> findByTrackid(int trackid);
 
 }
